requirements-dev.txt
pylint
pylint-odoo==3.0.3
bandit==1.6.2
